+++ api/celery.py
@@ -3,9 +3,12 @@ import os
 from celery import Celery
 from django.conf import settings
 
-app = Celery('api')
+app = Celery('api', include=[
+    "fondo_api.scheduler.tasks", 
+    "fondo_api.celery.tasks"
+])
 
 # Using a string here means the worker will not have to
 # pickle the object when using Windows.
 app.config_from_object('django.conf:settings')
-app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)
+# app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)
