+++ fondo_api/scheduler/tasks.py
@@ -1,33 +1,12 @@
 import logging
-import requests
-import json
-import os
 from celery.decorators import task, periodic_task
 from celery.task.schedules import crontab
 from datetime import datetime
 
-import fondo_api.logic.notifications_logic as notifications_logic
 from fondo_api.models import SchedulerTask
 from fondo_api.scheduler.executers.factory import get_executer
 
 logger = logging.getLogger(__name__)
-
-@task(name = "send_notification")
-def send_notification(message):
-    logger.info("Sending request to MNS...")
-    mns_host = os.environ.get('MNS_HOST', 'localhost')
-    try:
-        request = requests.post('http://{}:9901/mns/send'.format(mns_host), \
-            data = json.dumps(message['content']), \
-            headers = { 'content-type': 'application/json' }, \
-            timeout = 120 \
-        )
-        # remove invalid subscriptions
-        if request.status_code == requests.codes.ok:
-            invalid_subscriptions = request.json()['invalid']
-            notifications_logic.remove_invalid_subscriptions(invalid_subscriptions)
-    except Exception as ex:
-        logger.error("Error trying to connect to MNS service: {}".format(ex))
 
 @periodic_task(
     name = "Scheduler",
