+++ fondo_api/services/alexa/handlers/intent_handler.py
@@ -1,7 +1,7 @@
 import logging
 
-from fondo_api.logic.alexa.handlers.abstract_handler import AbstractRequestHandler
-from fondo_api.logic.alexa.intents.request_loan_intent import RequestLoanIntent
+from fondo_api.services.alexa.handlers.abstract_handler import AbstractRequestHandler
+from fondo_api.services.alexa.intents.request_loan_intent import RequestLoanIntent
 
 logger = logging.getLogger(__name__)
