+++ fondo_api/views/user.py
@@ -4,12 +4,16 @@ from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.parsers import MultiPartParser
 
-from fondo_api.logic.user_logic import *
+from fondo_api.services.user import UserService
+from fondo_api.services.notification import NotificationService
+
+notification_service = NotificationService()
+user_service = UserService(notification_service)
 
 class UserView(APIView):
 
     def post(self, request):
-        state, msg = create_user(request.data)
+        state, msg = user_service.create_user(request.data)
         if state:
             return Response(status=status.HTTP_201_CREATED)
         return Response({'message': msg}, status=status.HTTP_409_CONFLICT)
@@ -17,12 +21,15 @@ class UserView(APIView):
     def get(self, request):
         page = int(request.query_params.get('page', '1'))
         if page <= 0:
-            return Response({'message': 'Page number must be greater or equal than 0'}, status=status.HTTP_400_BAD_REQUEST)
-        return Response(get_users(page), status=status.HTTP_200_OK)
+            return Response(
+                {'message': 'Page number must be greater or equal than 0'}, 
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        return Response(user_service.get_users(page), status=status.HTTP_200_OK)
 
     @parser_classes((MultiPartParser,))
     def patch(self, request):
-        bulk_update_users(request.data)
+        user_service.bulk_update_users(request.data)
         return Response(status=status.HTTP_200_OK)
 
 class UserDetailView(APIView):
@@ -31,14 +38,14 @@ class UserDetailView(APIView):
         id = int(id)
         if id == -1:
             id = request.user.id
-        state, data = get_user(id)
+        state, data = user_service.get_user(id)
         if state:
             return Response(data, status=status.HTTP_200_OK)
         return Response(status=status.HTTP_404_NOT_FOUND)
 
     def patch(self, request, id):
         id = int(id)
-        state, code = update_user(id, request.data)
+        state, code = user_service.update_user(id, request.data)
         if state:
             return Response(status=status.HTTP_200_OK)
         elif code == 404:
@@ -47,7 +54,7 @@ class UserDetailView(APIView):
 
     def delete(self, request, id):
         id = int(id)
-        state = inactive_user(id)
+        state = user_service.inactive_user(id)
         if state:
             return Response(status=status.HTTP_200_OK)
         return Response(status=status.HTTP_404_NOT_FOUND)            
@@ -56,7 +63,7 @@ class UserActivateView(APIView):
     permission_classes = []
 
     def post(self, request, id):
-        state = activate_user(id, request.data)
+        state = user_service.activate_user(id, request.data)
         if state:
             return Response(status=status.HTTP_200_OK)
         return Response(status=status.HTTP_404_NOT_FOUND)
